+++ src/renderer/src/stores/track-store.svelte.ts
@@ -1,5 +1,5 @@
 import { deriveCommonMetadata } from '@domain/editor/batch-metadata';
-import { DEFAULT_GENRES } from '@domain/flac/types';
+import { DEFAULT_GENRES } from '@domain/flac/constants';
 import { createImageUrl } from '@renderer/utils/image';
 import { selectionState } from './selection-state.svelte';
 import { TrackRecord } from './track-record.svelte';
